+++ pyccel/parser/semantic.py
@@ -1094,7 +1094,7 @@ class SemanticParser(BasicParser):
         func : FunctionDef instance, Interface instance or PyccelInternalFunction type
                The function being called.
 
-        args : tuple
+        args : iterable
                The arguments passed to the function.
 
         is_method : bool
@@ -1118,10 +1118,13 @@ class SemanticParser(BasicParser):
 
             try:
                 new_expr = func(*args, **kwargs)
-            except TypeError:
-                errors.report(UNRECOGNISED_FUNCTION_CALL,
-                        symbol = expr,
-                        severity = 'fatal')
+            except TypeError as e:
+                message = str(e)
+                if not message:
+                    message = UNRECOGNISED_FUNCTION_CALL
+                errors.report(message,
+                                symbol = expr,
+                                severity = 'fatal')
 
             return new_expr
         else: